requirements.txt
python-dotenv=1.0.1
fastapi==0.115.4 
uvicorn==0.32.0
dependency-injector==4.43.0
sqlmodel==0.0.22
psycopg2-binary==2.9.10
pytz==2024.2